NagaAgent_v5.0.0_20260222_223727/NagaAgent_v5.0.0_20260222_223727/requirements_install.txt
# NagaAgent Runtime Dependencies
# Python 3.11+ recommended

# LLM / Chat Core
openai>=2.8.0
langchain-community>=0.3.30
langchain-openai>=1.1.0
litellm>=1.80.0
tiktoken>=0.11.0

# Core / Config
nagaagent-core>=1.0.9
pydantic>=2.11.0
pyyaml>=6.0.2
json5>=0.9.0
charset-normalizer>=3.4.0

# Web Service / API
fastapi>=0.115.0
uvicorn>=0.34.0
sse-starlette>=2.3.0
starlette
flask>=3.1.0
flask-socketio>=5.3.6
python-socketio>=5.8.0
eventlet>=0.33.3
gevent>=25.0.0
gevent-websocket
flask-cors
requests>=2.32.0

# Async / Task
aiohttp>=3.12.0
anyio
aiofiles>=23.2.1
apscheduler>=3.11.0

# Voice / Audio
pydub>=0.25.0
soundfile>=0.12.0
sounddevice
pyaudio>=0.2.11
simpleaudio
librosa>=0.11.0
mutagen
pygame>=2.6.1

# Browser / Automation
playwright>=1.48.0
selenium>=4.0.0

# Database / Graph
neo4j>=5.26.0
py2neo>=2021.2.4
pymysql>=1.1.0
aiomysql>=0.2.0
aiosqlite>=0.21.0
asyncmy>=0.2.10
redis>=4.6.0
SQLAlchemy>=2.0.35
asyncpg>=0.29.0
alembic>=1.16.5

# File Processing
openpyxl>=3.1.0
pandas>=2.2.0
Pillow>=11.0.0
numpy>=1.24.0
opencv-python>=4.8.0
opencv-contrib-python>=4.10.0.84
pytesseract>=0.3.10

# UI / GUI
PyQt5>=5.15.0
PyQtWebEngine>=5.15.0
PyQt5-Qt5
PyQt5_sip
streamlit>=1.28.1

# Desktop Automation (baodou_AI)
PyAutoGUI>=0.9.54
PyGetWindow>=0.0.9
PyRect>=0.2.0
PyScreeze>=1.0.1
pytweening>=1.2.0
MouseInfo>=0.1.3
PyMsgBox>=2.0.1
pywin32>=306
pywinauto>=0.6.9
comtypes>=1.4.11

# Logging / Utils
colorlog>=6.8.0
python-dotenv>=1.0.0
python-dateutil>=2.8.2
pytz>=2023.3
tqdm>=4.65.0
tenacity>=8.2.2
loguru>=0.7.0

# MCP
httpx>=0.27.0
pydantic-settings>=2.6.0
fastmcp>=2.14.4
crawl4ai>=0.3.0
mcp>=1.8.1
websockets>=14.0

# External Tools
pyperclip>=1.8.0
psutil>=5.9.0
chardet>=5.2.0
pathlib2>=2.3.7

# Search / Web
tavily-python>=0.3.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
parsel>=1.9.1
pyexecjs>=1.5.1

# Text Processing
regex>=2023.8.8
jieba>=0.42.1
pypinyin>=0.53.0
gensim>=4.0.1

# Markdown / Docs
markdown>=3.10
types-markdown>=3.10.0.20251106
markdown-it-py[plugins]>=4.0.0
mdit-py-plugins>=0.5.0
python-markdown-math>=0.9
linkify-it-py>=2.0.3
pymdown-extensions>=10.20

# Document Processing
pymupdf>=1.26.7
python-docx>=1.2.0
python-pptx>=1.0.2
py7zr>=1.1.0
rarfile>=4.2
docx2pdf
msoffcrypto-tool

# Visualization
plotly>=5.17.0
matplotlib>=3.9.0
wordcloud>=1.9.3

# Astronomy / Math
sympy
scipy
skyfield

# Chinese Calendar
lunar-python>=1.4.8

# Output / Rendering
imgkit

# Console UI
rich>=14.2.0

# Task Scheduling
croniter>=2.0.0

# Live2D
live2d-py<0.6

# MQTT
paho-mqtt>=2.1.0

# Computer Vision / AI
onnxruntime>=1.23.2
onnx>=1.20.0

# Tray Icon
pystray>=0.19.5

# GUI-Agents
gui-agents

# Cloud Storage
cos-python-sdk-v5>=1.9.38

# WeChat / QQ
itchat>=1.2.0

# String Matching
rapidfuzz

# Machine Learning (Optional - may fail, installation continues)
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0
xgboost>=2.0.0
peft>=0.4.0
datasets>=2.0.0
accelerate>=0.20.0
sentence-transformers>=2.7.0
bitsandbytes>=0.41.0
safetensors>=0.3.0

# Web Frameworks
django>=5.2.9
twisted>=25.5.0

# AI Models
dashscope>=1.25.3
bilibili-api>=4.1.0

# CLI Tools
typer>=0.12.3
pyhumps>=3.8.0

# Network / Crypto
curl-cffi
pycryptodome
cryptography>=45.0.7
urllib3>=2.6.2

# Misc
commonx>=0.6.38
zhconv
annotated-types>=0.7.0
certifi>=2025.11.12
distro>=1.9.0
h11>=0.16.0
httpcore>=1.0.9
idna>=3.11
jiter>=0.12.0
packaging>=25.0
pefile>=2024.8.26
pywin32-ctypes>=0.2.3
sniffio>=1.3.1
typing-inspection>=0.4.2

# Build Tools
pyinstaller
pyinstaller-hooks-contrib

# Testing / Dev (Optional)
pytest>=7.0.0
pytest-asyncio>=0.20.0
black>=22.0.0
ruff>=0.1.0
mypy>=1.8.0
expecttest>=0.3.0

# Machine Learning Frameworks (Optional)
tensorflow>=2.20.0
tensorboard>=2.20.0

# 3D Graphics
pyopengl>=3.1.10
pyopengl-accelerate>=3.1.10

# Windows Specific
win10toast

# MCP-O
mcpo
